+++ Base/Market.cs
@@ -38,22 +38,22 @@ namespace AgentBit.Ccxt.Base
 
         public bool Active { get; set; }
         public int PricePrecision { get; set; }
-        public double PriceMin { get; set; }
-        public double PriceMax { get; set; }
+        public decimal PriceMin { get; set; }
+        public decimal PriceMax { get; set; }
         public int AmountPrecision { get; set; }
-        public double AmountMin { get; set; }
-        public double AmountMax { get; set; }
-        public double CostMin { get; set; }
-        public double CostMax { get; set; }
+        public decimal AmountMin { get; set; }
+        public decimal AmountMax { get; set; }
+        public decimal CostMin { get; set; }
+        public decimal CostMax { get; set; }
 
         /// <summary>
         /// Default max maker fee rate, 0.0016 = 0.16%
         /// </summary>
-        public double FeeMaker { get; set; }
+        public decimal FeeMaker { get; set; }
         /// <summary>
         /// Default max taker fee rate, 0.002 = 0.2%
         /// </summary>
-        public double FeeTaker { get; set; }
+        public decimal FeeTaker { get; set; }
 
         public object Info { get; set; }
         public string Url { get; set; }
@@ -61,9 +61,9 @@ namespace AgentBit.Ccxt.Base
         public Market()
         {
             Active = true;
-            PriceMax = double.MaxValue;
-            AmountMax = double.MaxValue;
-            CostMax = double.MaxValue;
+            PriceMax = decimal.MaxValue;
+            AmountMax = decimal.MaxValue;
+            CostMax = decimal.MaxValue;
         }
     }
 }
